chore(deps): simplify dep management across dev environments
Make it easier to manage dependencies across multiple dev environments,
while maintaining the original setup_venv script functionality.

Dependency installation and setup is extracted to scripts/setup-deps.sh,
and the setup_venv script has been refactored to use it. This split
allows us to use venv, virtualenv, or Docker with the docs-builder.

Instead of using setuputils, we use pip to install the local plugins in
development mode (safer alternative to 'setup.py develop') which is
equivalent but is able to manage the dependencies with pip.

Calls to the 'pip' binary has been replaced with 'pip3' to ensure that
we use a python3 compatible pip version.

"Temporary hacks" may not be necessary when using pip3 to install the
deps, but has been maintained because I cannot verify this locally on
account of not being able to use venv.

Finally the README has been updated with additional prerequisites for
docs-builder to work (cairo and pango).